litmodel3d/demo/requirements.txt
gradio_litmodel3d